rpsttokenexchange/requirements.txt
fdk>=0.1.103
oci>=2.112.0
requests>=2.25.1
google-auth>=2.0.0
google-cloud-storage>=2.0.0